requirements.txt
bleach==1.5.0
bottle==0.12.13
click==6.7
enum34==1.1.6
h5py==2.7.1
html5lib==0.9999999
Keras==2.1.2
Markdown==2.6.10
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
protobuf==3.5.0.post1
python-dotenv==0.7.1
PyYAML==3.12
scipy==1.0.0
six==1.11.0
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
Werkzeug==0.13
